Clean up new framework stuff
