+++ src/osp/tasks/tasks.cpp
@@ -26,7 +26,7 @@
 
 #include <longeron/id_management/id_set_stl.hpp>
 
-#include <array>
+#include <sstream>
 
 namespace osp
 {
@@ -44,14 +44,10 @@ struct StageLock
     StageId     stageId;
 };
 
-void check_task_order(Tasks const& tasks, TaskOrderReport& rOut, lgrn::IdSetStl<LoopBlockId> const& loopblks)
+void check_task_order(Tasks const& tasks, TaskOrderReport& rOut, lgrn::IdSetStl<LoopBlockId> loopblks)
 {
-
-    // step 1: add all the tasks
-
     std::vector<StageLock> locks;
 
-    // add some way to know which stages are connected to which tasks
     osp::KeyedVec<PipelineId, StageId> currentStages;
     currentStages.resize(tasks.pipelineIds.capacity());
     lgrn::IdSetStl<PipelineId> stageJustChanged;
@@ -91,34 +87,33 @@ void check_task_order(Tasks const& tasks, TaskOrderReport& rOut, lgrn::IdSetStl<
 
     auto &rPltypeinfo = PipelineTypeIdReg::instance();
 
-    while (true)
+    auto const task_aligned = [&] (TaskId const taskId, PipelineId const pipelineId, StageId stageId)
     {
-        // consider connected points
-
-        auto const task_aligned = [&] (TaskId const taskId, PipelineId const pipelineId, StageId stageId)
+        LGRN_ASSERT(alignsRemaining[taskId] != 0);
+        -- alignsRemaining[taskId];
+        if (alignsRemaining[taskId] == 0)
         {
-            LGRN_ASSERT(alignsRemaining[taskId] != 0);
-            -- alignsRemaining[taskId];
-            if (alignsRemaining[taskId] == 0)
-            {
-                std::cout << "task passed: " << taskId.value << "\n";
-                rOut.steps.push_back({.taskId = taskId, .time = time});
+            rOut.steps.push_back({.taskId = taskId, .time = time});
 
-                // remove all locks assiciated with this task
-                for (StageLock &rLock : locks)
+            // remove all locks assiciated with this task
+            for (StageLock &rLock : locks)
+            {
+                if (rLock.taskId == taskId)
                 {
-                    if (rLock.taskId == taskId)
-                    {
-                        rLock = {};
-                    }
+                    rLock = {};
                 }
             }
-            else
-            {
-                locks.push_back({taskId, pipelineId, stageId});
-            }
-        };
+        }
+        else
+        {
+            locks.push_back({taskId, pipelineId, stageId});
+        }
+    };
 
+    bool nothingMoved = true;
+
+    do
+    {
         // any points connected to current stages?
         for (TaskSyncToPipeline const& syncTo : tasks.syncs)
         {
@@ -146,11 +141,7 @@ void check_task_order(Tasks const& tasks, TaskOrderReport& rOut, lgrn::IdSetStl<
         auto const &locksFirst = locks.begin();
         auto const &locksLast  = locks.end();
 
-
-
         stageJustChanged.clear();
-
-        bool nothingMoved = true;
 
         // move points
         for (PipelineId const pipelineId : tasks.pipelineIds)
@@ -190,12 +181,7 @@ void check_task_order(Tasks const& tasks, TaskOrderReport& rOut, lgrn::IdSetStl<
             }
         }
 
-        if (nothingMoved)
-        {
-            break;
-        }
-
-    }
+    } while (!nothingMoved);
 
     for (TaskId const taskId : tasks.taskIds)
     {
@@ -213,6 +199,186 @@ void check_task_order(Tasks const& tasks, TaskOrderReport& rOut, lgrn::IdSetStl<
             }
         }
     }
+
+    rOut.loopblks = std::move(loopblks);
+}
+
+std::string visualize_task_order(TaskOrderReport const& report, Tasks const& tasks)
+{
+    std::ostringstream os;
+    auto        it    = report.steps.begin();
+    auto const  &last = report.steps.end();
+    int         time  = 0;
+
+    int count = 0;
+    for(PipelineId const pipelineId : tasks.pipelineIds)
+    {
+        Pipeline const& pipeline = tasks.pipelineInst[pipelineId];
+        if (report.loopblks.contains(pipeline.block))
+        {
+            for (int i = 0; i < count; ++i)
+            {
+                os << "\U00002502 ";
+            }
+
+            os << "\U0000250D" << pipeline.name << "\n";
+            ++count;
+        }
+
+    }
+
+    osp::KeyedVec<PipelineId, StageId> stageChanges;
+
+    while(it != last)
+    {
+        stageChanges.clear();
+        stageChanges.resize(tasks.pipelineIds.capacity());
+        while(it != last && it->pipelineId.has_value() && it->time == time)
+        {
+            stageChanges[it->pipelineId] = it->stageId;
+            ++it;
+        }
+
+        for(PipelineId const pipelineId : tasks.pipelineIds)
+        {
+            if (report.loopblks.contains(tasks.pipelineInst[pipelineId].block))
+            {
+                StageId const stageChange = stageChanges[pipelineId];
+                if (stageChange.has_value())
+                {
+                    os << int(stageChange.value) << " ";
+                }
+                else
+                {
+                    os << "\U00002502 ";
+                }
+            }
+
+        }
+
+        os << "\n";
+
+        while(it != last && it->taskId.has_value() && it->time == time)
+        {
+            for(PipelineId const pipelineId : tasks.pipelineIds)
+            {
+                if (report.loopblks.contains(tasks.pipelineInst[pipelineId].block))
+                {
+                    if (tasks.pipelineInst[pipelineId].scheduleCondition == it->taskId)
+                    {
+                        os << "\U000025C9\U00002500";
+                    }
+                    else if (std::find_if(
+                                    tasks.syncs.begin(), tasks.syncs.end(),
+                                    [taskId = it->taskId, pipelineId] (TaskSyncToPipeline const& sync)
+                                    { return sync.task == taskId && sync.pipeline == pipelineId; })
+                             != tasks.syncs.end())
+                    {
+                        os << "\U000025CF\U00002500";
+                    }
+                    else
+                    {
+                        os << "\U0000253C\U00002500";
+                    }
+                }
+            }
+            os << "" <<  tasks.taskInst[it->taskId].debugName << "\n";
+            ++it;
+        }
+
+        ++time;
+
+    }
+
+    bool noError = report.failedLocked.empty() && report.failedNotAdded.empty();
+
+    if ( ! noError )
+    {
+        auto  const &pltypereg  = PipelineTypeIdReg::instance();
+
+        os << "Deadlocked tasks:\n";
+        for (TaskId const taskId : report.failedLocked)
+        {
+            for(PipelineId const pipelineId : tasks.pipelineIds)
+            {
+                if (report.loopblks.contains(tasks.pipelineInst[pipelineId].block))
+                {
+                    Pipeline const& pipeline = tasks.pipelineInst[pipelineId];
+                    if (pipeline.scheduleCondition == taskId)
+                    {
+                        auto const &pltype = pltypereg.get(pipeline.type);
+                        auto stageIndex = std::distance(
+                                std::find_if(pltype.stages.begin(), pltype.stages.end(),
+                                             [] (PipelineTypeInfo::StageInfo const &stage) { return stage.isSchedule; } ),
+                                pltype.stages.end());
+                        os << stageIndex << "\U00002500";
+                    }
+                    else if (auto const syncIt = std::find_if(
+                                    tasks.syncs.begin(), tasks.syncs.end(),
+                                    [taskId, pipelineId] (TaskSyncToPipeline const& sync)
+                                    { return sync.task == taskId && sync.pipeline == pipelineId; });
+                            syncIt != tasks.syncs.end())
+                    {
+                        StageId const lastStage = std::find_if(
+                                report.steps.rbegin(), report.steps.rend(),
+                                [pipelineId] (TaskOrderReport::Step const& step)
+                                { return step.pipelineId == pipelineId; })->stageId;
+                        if (lastStage == syncIt->stage)
+                        {
+                            os << "\U000025CF\U00002500";
+                        }
+                        else
+                        {
+                            os << int(syncIt->stage.value) << "\U00002500";
+                        }
+                    }
+                    else
+                    {
+                        os << "\U00002534\U00002500";
+                    }
+                }
+            }
+            os << tasks.taskInst[taskId].debugName << "\n";
+        }
+
+        os << "Not reached:\n";
+
+        for (TaskId const taskId : report.failedNotAdded)
+        {
+            for(PipelineId const pipelineId : tasks.pipelineIds)
+            {
+                if (report.loopblks.contains(tasks.pipelineInst[pipelineId].block))
+                {
+                    Pipeline const& pipeline = tasks.pipelineInst[pipelineId];
+                    if (pipeline.scheduleCondition == taskId)
+                    {
+                        auto const &pltype = pltypereg.get(pipeline.type);
+                        auto stageIndex = std::distance(
+                                std::find_if(pltype.stages.begin(), pltype.stages.end(),
+                                             [] (PipelineTypeInfo::StageInfo const &stage) { return stage.isSchedule; } ),
+                                pltype.stages.end());
+                        os << stageIndex << "\U00002500";
+                    }
+                    else if (auto const syncIt = std::find_if(
+                                    tasks.syncs.begin(), tasks.syncs.end(),
+                                    [taskId, pipelineId] (TaskSyncToPipeline const& sync)
+                                    { return sync.task == taskId && sync.pipeline == pipelineId; });
+                            syncIt != tasks.syncs.end())
+                    {
+                        os << int(syncIt->stage.value) << "\U00002500";
+                    }
+                    else
+                    {
+                        os << "\U00002534\U00002500";
+                    }
+                }
+            }
+            os << "" <<  tasks.taskInst[taskId].debugName << "\n";
+        }
+
+    }
+
+    return os.str();
 }
 
 } // namespace osp
