+++ test/sync_graph/CMakeLists.txt
@@ -26,4 +26,7 @@ PROJECT(test_sync_graph CXX)
 ADD_TEST_DIRECTORY(${PROJECT_NAME})
 
 TARGET_LINK_LIBRARIES(${PROJECT_NAME} PRIVATE longeron EnTT::EnTT Magnum::Magnum)
-TARGET_SOURCES(${PROJECT_NAME} PRIVATE "${CMAKE_SOURCE_DIR}/src/osp/executor/sync_graph.cpp")
+TARGET_SOURCES(${PROJECT_NAME} PRIVATE
+    "${CMAKE_SOURCE_DIR}/src/osp/executor/sync_graph.cpp"
+    "${CMAKE_SOURCE_DIR}/src/osp/executor/singlethread_sync_graph.cpp"
+)
